refactored Shim expression traits in 332db
